+++ src/ClearingHouseViewer.sol
@@ -3,13 +3,14 @@ pragma solidity 0.6.9;
 pragma experimental ABIEncoderV2;
 
 import { IERC20 } from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
-import { Decimal } from "./utils/Decimal.sol";
+import { SafeMath, Decimal } from "./utils/Decimal.sol";
 import { SignedDecimal, MixedDecimal } from "./utils/MixedDecimal.sol";
 import { IAmm } from "./interface/IAmm.sol";
 import { IInsuranceFund } from "./interface/IInsuranceFund.sol";
 import { ClearingHouse } from "./ClearingHouse.sol";
 
 contract ClearingHouseViewer {
+    using SafeMath for uint256;
     using Decimal for Decimal.decimal;
     using SignedDecimal for SignedDecimal.signedDecimal;
     using MixedDecimal for SignedDecimal.signedDecimal;
@@ -84,6 +85,24 @@ contract ClearingHouseViewer {
             getFundingPayment(position, clearingHouse.getLatestCumulativePremiumFraction(_amm))
         );
         position.margin = marginWithFundingPayment.toInt() >= 0 ? marginWithFundingPayment.abs() : Decimal.zero();
+    }
+
+    /**
+     * @notice verify if trader's position needs to be migrated
+     * @param _amm IAmm address
+     * @param _trader trader address
+     * @return true if trader's position is not at the latest Amm curve, otherwise is false
+     */
+    function isPositionNeedToBeMigrated(IAmm _amm, address _trader) external view returns (bool) {
+        ClearingHouse.Position memory unadjustedPosition = clearingHouse.getUnadjustedPosition(_amm, _trader);
+        if (unadjustedPosition.size.toInt() == 0) {
+            return false;
+        }
+        uint256 latestLiquidityIndex = _amm.getLiquidityHistoryLength().sub(1);
+        if (unadjustedPosition.liquidityHistoryIndex == latestLiquidityIndex) {
+            return false;
+        }
+        return true;
     }
 
     /**